+++ src/eval.hpp
@@ -5,4 +5,4 @@
 
 // Tapered static evaluation function given a board position
 // returns score relative to player
-int16_t evaluate(const chess::Board& board);
+int32_t evaluate(const chess::Board& board);
